* project documentation initialization * several code clean-ups * cutting planes and bundle methods tests now automatically skip if gurobipy is not installed * improved README * simplified requirements.txt
