+++ .gitignore
@@ -14,3 +14,5 @@ nsopy.egg-info
 
 dist
 build
+
+venv/
