+++ .readthedocs.yaml
@@ -1,0 +1,18 @@
+version: 2
+
+build:
+  os: "ubuntu-20.04"
+  tools:
+    python: "3.8"
+
+python:
+  install:
+    - method: pip
+      path: .
+
+sphinx:
+  fail_on_warning: true
+
+formats:
+  - pdf
+  - epub
