+++ nsopy/methods/universal.py
@@ -1,4 +1,3 @@
-from __future__ import division
 import numpy as np
 import copy
