+++ requirements.txt
@@ -1,8 +1,2 @@
-colorama>=0.3.7
 numpy>=1.11.2
 pandas>=0.19.0
-py>=1.4.34
-pytest>=3.0.3
-python-dateutil>=2.5.3
-pytz>=2016.7
-six>=1.10.0
