+++ tests/test_subgradient_methods.py
@@ -1,5 +1,3 @@
-from __future__ import print_function
-
 import time
 
 import numpy as np
